pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "frederick-population-pipeline"
version = "0.1.0"
description = "Daily Frederick, Maryland public-information population pipeline"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "feedparser>=6.0.11",
  "python-dateutil>=2.9.0.post0",
  "requests>=2.32.3",
  "trafilatura>=1.9.0"
]

[project.scripts]
frederick-pipeline = "frederick_pipeline.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
